Migrate tokens-alias files to win32-theme
