--- packages/theming/theme-tokens/src/black/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '@fluentui-react-native/design-tokens-win32/black/tokens-aliases.json';
-
-export default aliasTokens;
