--- packages/theming/theme-tokens/src/colorful/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '@fluentui-react-native/design-tokens-win32/colorful/tokens-aliases.json';
-
-export default aliasTokens;
