--- packages/theming/theme-tokens/src/darkGray/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '@fluentui-react-native/design-tokens-win32/darkgray/tokens-aliases.json';
-
-export default aliasTokens;
