+++ scripts/build-tokens.js
@@ -26,13 +26,6 @@ async function run() {
   fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/global/reactnative/tokens-aliases.json'));
   fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/global/reactnative/tokens-controls.json'));
 
-  console.log('Generating win32 global tokens...');
-  child_process.execSync(
-    'yarn transform-tokens --in ./packages/theming/theme-tokens/src/pipeline-input/token-input.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-brand.win32.json --out ./packages/theming/theme-tokens/src/generated/global-win32 --p reactnative',
-  );
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/global-win32/reactnative/tokens-aliases.json'));
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/global-win32/reactnative/tokens-controls.json'));
-
   console.log('Generating macOS light mode global and alias  tokens...');
   child_process.execSync(
     'yarn transform-tokens --in ./packages/theming/theme-tokens/src/pipeline-input/token-input.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-global.macos.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-brand-light.macos.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-light.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-light.macos.json --out ./packages/theming/theme-tokens/src/generated/light-macos --p reactnative',
@@ -88,27 +81,6 @@ async function run() {
   );
   fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/dark/reactnative/tokens-global.json'));
   fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/dark/reactnative/tokens-controls.json'));
-
-  console.log('Generating colorful mode tokens...');
-  child_process.execSync(
-    'yarn transform-tokens --in ./packages/theming/theme-tokens/src/pipeline-input/token-input.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-brand.win32.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-light.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-colorful.json --out ./packages/theming/theme-tokens/src/generated/colorful --p reactnative',
-  );
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/colorful/reactnative/tokens-global.json'));
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/colorful/reactnative/tokens-controls.json'));
-
-  console.log('Generating dark gray mode tokens...');
-  child_process.execSync(
-    'yarn transform-tokens --in ./packages/theming/theme-tokens/src/pipeline-input/token-input.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-brand.win32.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-darkgray.json --out ./packages/theming/theme-tokens/src/generated/darkGray --p reactnative',
-  );
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/darkGray/reactnative/tokens-global.json'));
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/darkGray/reactnative/tokens-controls.json'));
-
-  console.log('Generating black mode tokens...');
-  child_process.execSync(
-    'yarn transform-tokens --in ./packages/theming/theme-tokens/src/pipeline-input/token-input.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-brand.win32.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-dark.json --in ./packages/theming/theme-tokens/src/pipeline-input/token-input-black.json --out ./packages/theming/theme-tokens/src/generated/black --p reactnative',
-  );
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/black/reactnative/tokens-global.json'));
-  fs.unlinkSync(path.join(process.cwd(), 'packages/theming/theme-tokens/src/generated/black/reactnative/tokens-controls.json'));
 
   console.log('Generating high contrast mode tokens...');
   child_process.execSync(
